Use code-inline inside code-block
