+++ addon/components/code-block.hbs
@@ -1,14 +1,11 @@
 <pre
-  ...attributes
-  class="code-block {{this.languageClass}} {{if @showLineNumbers "line-numbers"}}"
+  class="{{this.languageClass}} {{if @showLineNumbers "line-numbers"}}"
 >
   {{~! ~}}
-  <code
-    class="{{this.languageClass}}"
-    {{did-insert this.setPrismCode}}
-    {{did-update this.setPrismCode this.code @language}}
-  >
-    {{~! ~}}{{this.prismCode}}{{~! ~}}
-  </code>
+  <CodeInline
+    ...attributes
+    @code={{@code}}
+    @language={{@language}}
+  />
   {{~! ~}}
 </pre>
