+++ addon/components/code-block.js
@@ -1,4 +1,11 @@
-import CodeBaseComponent from './code-base';
+import Component from '@glimmer/component';
 
-export default class NewCodeBlockComponent extends CodeBaseComponent {
+export default class NewCodeBlockComponent extends Component {
+  get language() {
+    return this.args.language ?? 'markup';
+  }
+
+  get languageClass() {
+    return `language-${this.language}`;
+  }
 }
