+++ addon/components/code-inline.hbs
@@ -1,6 +1,6 @@
 <code
   ...attributes
-  class="code-inline {{this.languageClass}}"
+  class="{{this.languageClass}}"
   {{did-insert this.setPrismCode}}
   {{did-update this.setPrismCode this.code @language}}
 >
