+++ tests/acceptance/code-block-test.js
@@ -9,13 +9,13 @@ module('Acceptance | code block', function(hooks) {
   test('has `line-numbers` plugin', async function(assert) {
     await visit('/');
 
-    assert.dom('pre.code-block.line-numbers').exists('`pre.codeblock.line-numbers` is added to the template');
-    assert.dom('pre.code-block .line-numbers-rows').exists('`line-numbers` plugin generates `.line-numbers-rows`');
+    assert.dom('pre.line-numbers').exists('`pre.line-numbers` is added to the template');
+    assert.dom('pre .line-numbers-rows').exists('`line-numbers` plugin generates `.line-numbers-rows`');
   });
 
   test('the pre innerHTML is trimmed when rendered', async function(assert) {
     await visit('/');
 
-    assert.equal(find('pre.can-you-dig-it').innerHTML.split('\n').length, 3);
+    assert.equal(find('pre code.can-you-dig-it').innerHTML.split('\n').length, 3);
   });
 });
